requirements.txt
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
torchviz==0.0.2
tqdm==4.62.3
scikit-image==0.19.2
scikit-learn==1.0.2
seaborn==0.11.2
pillow==8.4.0
pandas==1.3.4
opencv-python==4.5.5.62
numpy==1.20.3
natsort==8.0.0
matplotlib==3.5.1
btrack==0.4.3
aicspylibczi==3.0.5